refactor to replace v1 by v2 and cleanup suffix
