+++ multilinear_extensions/src/lib.rs
@@ -2,7 +2,6 @@
 pub mod mle;
 pub mod util;
 pub mod virtual_poly;
-pub mod virtual_poly_v2;
 
 #[cfg(test)]
 mod test;
