+++ multilinear_extensions/src/test.rs
@@ -32,31 +32,6 @@ fn test_virtual_polynomial_additions() {
 }
 
 #[test]
-fn test_virtual_polynomial_mul_by_mle() {
-    let mut rng = test_rng();
-    for nv in 2..5 {
-        for num_products in 2..5 {
-            let base: Vec<E> = (0..nv).map(|_| E::random(&mut rng)).collect();
-
-            let (a, _a_sum) = VirtualPolynomial::<E>::random(nv, (2, 3), num_products, &mut rng);
-            let (b, _b_sum) = DenseMultilinearExtension::<E>::random_mle_list(nv, 1, &mut rng);
-            let b_mle = b[0].clone();
-            let coeff = Goldilocks::random(&mut rng);
-            let b_vp = VirtualPolynomial::new_from_mle(b_mle.clone(), coeff);
-
-            let mut c = a.clone();
-
-            c.mul_by_mle(b_mle, coeff);
-
-            assert_eq!(
-                a.evaluate(base.as_ref()) * b_vp.evaluate(base.as_ref()),
-                c.evaluate(base.as_ref())
-            );
-        }
-    }
-}
-
-#[test]
 fn test_eq_xr() {
     let mut rng = test_rng();
     for nv in 4..10 {
