+++ sumcheck/src/lib.rs
@@ -2,7 +2,6 @@
 #![feature(decl_macro)]
 pub mod macros;
 mod prover;
-mod prover_v2;
 pub mod structs;
 pub mod util;
 mod verifier;
