+++ sumcheck/src/util.rs
@@ -2,7 +2,6 @@ use std::{
     array,
     cmp::max,
     iter::Sum,
-    mem,
     ops::{Add, AddAssign, Deref, DerefMut, Mul, MulAssign},
     sync::Arc,
 };
@@ -11,14 +10,11 @@ use ark_std::{end_timer, start_timer};
 use ff::PrimeField;
 use ff_ext::ExtensionField;
 use multilinear_extensions::{
-    mle::{DenseMultilinearExtension, FieldType},
-    op_mle,
-    virtual_poly::VirtualPolynomial,
-    virtual_poly_v2::VirtualPolynomialV2,
+    mle::DenseMultilinearExtension, op_mle, virtual_poly::VirtualPolynomial,
 };
 use rayon::{prelude::ParallelIterator, slice::ParallelSliceMut};
 
-use crate::structs::{IOPProverState, IOPProverStateV2};
+use crate::structs::IOPProverState;
 
 pub fn barycentric_weights<F: PrimeField>(points: &[F]) -> Vec<F> {
     let mut weights = points
@@ -221,41 +217,10 @@ pub fn ceil_log2(x: usize) -> usize {
     usize_bits - (x - 1).leading_zeros() as usize
 }
 
-pub(crate) fn merge_sumcheck_polys<E: ExtensionField>(
-    prover_states: &[IOPProverState<E>],
+pub(crate) fn merge_sumcheck_polys<'a, E: ExtensionField>(
+    prover_states: &[IOPProverState<'a, E>],
     max_thread_id: usize,
-) -> VirtualPolynomial<E> {
-    let log2_max_thread_id = ceil_log2(max_thread_id);
-    let mut poly = prover_states[0].poly.clone(); // giving only one evaluation left, this clone is low cost.
-    poly.aux_info.num_variables = log2_max_thread_id; // size_log2 variates sumcheck
-    for i in 0..poly.flattened_ml_extensions.len() {
-        let ml_ext = Arc::make_mut(&mut poly.flattened_ml_extensions[i]);
-        let _ = mem::replace(&mut ml_ext.evaluations, {
-            let evaluations = prover_states
-                .iter()
-                .map(|prover_state| {
-                    if let FieldType::Ext(evaluations) =
-                        &prover_state.poly.flattened_ml_extensions[i].evaluations
-                    {
-                        assert!(evaluations.len() == 1);
-                        evaluations[0]
-                    } else {
-                        unreachable!()
-                    }
-                })
-                .collect::<Vec<E>>();
-            assert!(evaluations.len() == max_thread_id);
-            FieldType::Ext(evaluations)
-        });
-        ml_ext.num_vars = log2_max_thread_id;
-    }
-    poly
-}
-
-pub(crate) fn merge_sumcheck_polys_v2<'a, E: ExtensionField>(
-    prover_states: &[IOPProverStateV2<'a, E>],
-    max_thread_id: usize,
-) -> VirtualPolynomialV2<'a, E> {
+) -> VirtualPolynomial<'a, E> {
     let log2_max_thread_id = ceil_log2(max_thread_id);
     let mut poly = prover_states[0].poly.clone(); // giving only one evaluation left, this clone is low cost.
     poly.aux_info.max_num_variables = log2_max_thread_id; // size_log2 variates sumcheck
